Move source.h code to .cpp
